requirements.txt
google-cloud-documentai
google-cloud-storage
google-cloud-documentai-toolbox
google-cloud-vision
google-auth
google-auth-oauthlib
google-auth-httplib2
opencv-python
Pillow
pandas
numpy
pytesseract
pdf2image
pymupdf
fastapi
uvicorn
python-multipart
flask
werkzeug
pydantic
python-dotenv
loguru
tqdm
matplotlib
seaborn
plotly
streamlit 